requirement.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
pillow>=10.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiohttp>=3.9.0
